Run clang-format
