+++ march_hardware/include/march_hardware/EtherCAT/EthercatIO.h
@@ -28,8 +28,7 @@ typedef struct packed_bit64
   unsigned char b5;
   unsigned char b6;
   unsigned char b7;
-}
-packed_bit64;
+} packed_bit64;
 
 // union used for int64 and uint64 in combination with the above struct
 typedef union bit64
@@ -37,8 +36,7 @@ typedef union bit64
   int64 i;
   uint64 ui;
   packed_bit64 p;
-}
-bit64;
+} bit64;
 
 // struct used to send the x2 variable to the servo drive in the format that is required
 typedef struct packed_sd_x2
@@ -46,8 +44,7 @@ typedef struct packed_sd_x2
   uint16 time;
   unsigned char reserved;  // empty but reserved space
   uint8 integrityCounter;  // represents a 7bit int where the last bit is also part of reserved
-}
-packed_sd_x2;
+} packed_sd_x2;
 
 // struct used to easily get specific bytes from a 32 bit variable
 typedef struct packed_bit32
@@ -56,8 +53,7 @@ typedef struct packed_bit32
   unsigned char b1;
   unsigned char b2;
   unsigned char b3;
-}
-packed_bit32;
+} packed_bit32;
 
 // union used for int32, uint32, float and the x2 variable for the servo drive (sd) in combination with the above struct
 typedef union bit32
@@ -67,16 +63,14 @@ typedef union bit32
   float f;
   packed_sd_x2 ppt;
   packed_bit32 p;
-}
-bit32;
+} bit32;
 
 // struct used to easily get specific bytes from a 16 bit variable
 typedef struct packed_bit16
 {
   unsigned char b0;
   unsigned char b1;
-}
-packed_bit16;
+} packed_bit16;
 
 // union used for int16 and uint16 in combination with the above struct
 typedef union bit16
@@ -84,8 +78,7 @@ typedef union bit16
   int16 i;
   uint16 ui;
   packed_bit16 p;
-}
-bit16;
+} bit16;
 
 // union used for int8 and uint8 in combination with the above struct, unsigned char is used to read single byte
 // unbiased
@@ -94,8 +87,7 @@ typedef union bit8
   int8 i;
   uint8 ui;
   unsigned char b0;
-}
-bit8;
+} bit8;
 
 // functions to write #bit amounts of data to the EtherCAT train.
 // Takes the slave number where 0 is the master and the first slave is 1.
