+++ march_hardware/src/EtherCAT/EthercatIO.cpp
@@ -3,7 +3,6 @@
 
 namespace march4cpp
 {
-
 // TODO(Isha, Martijn, Tim) refactor this with more generic types
 union bit64 get_input_bit64(uint16 slave_no, uint8 module_index)
 {
