+++ march_hardware/src/IMotionCube.cpp
@@ -466,7 +466,7 @@ bool IMotionCube::goToTargetState(IMotionCubeTargetState targetState)
   {
     this->setControlWord(targetState.getControlWord());
     ROS_INFO_DELAYED_THROTTLE(2, "\tWaiting for '%s': %s", targetState.getDescription().c_str(),
-                      std::bitset<16>(this->getStatusWord()).to_string().c_str());
+                              std::bitset<16>(this->getStatusWord()).to_string().c_str());
     if (targetState.getState() == IMotionCubeTargetState::OPERATION_ENABLED.getState() &&
         this->getState(this->getStatusWord()) == IMCState::fault)
     {
