+++ march_hardware/test/TestPDOmap.cpp
@@ -35,11 +35,11 @@ TEST_F(PDOTest, multipleAddObjects)
 
 TEST_F(PDOTest, ObjectCounts)
 {
-    march4cpp::PDOmap pdoMapMISO = march4cpp::PDOmap();
-    pdoMapMISO.addObject(march4cpp::IMCObjectName::CurrentLimit);
-    std::map<march4cpp::IMCObjectName, int> misoByteOffsets = pdoMapMISO.map(1, march4cpp::dataDirection::miso);
-    ASSERT_EQ(1, misoByteOffsets.count(march4cpp::IMCObjectName::CurrentLimit));
-    ASSERT_EQ(0, misoByteOffsets.count(march4cpp::IMCObjectName::DCLinkVoltage));
+  march4cpp::PDOmap pdoMapMISO = march4cpp::PDOmap();
+  pdoMapMISO.addObject(march4cpp::IMCObjectName::CurrentLimit);
+  std::map<march4cpp::IMCObjectName, int> misoByteOffsets = pdoMapMISO.map(1, march4cpp::dataDirection::miso);
+  ASSERT_EQ(1, misoByteOffsets.count(march4cpp::IMCObjectName::CurrentLimit));
+  ASSERT_EQ(0, misoByteOffsets.count(march4cpp::IMCObjectName::DCLinkVoltage));
 }
 
 TEST_F(PDOTest, exceedMaxSize)
