+++ march_hardware_builder/include/march_hardware_builder/AllowedRobot.h
@@ -26,7 +26,7 @@ public:
     }
     else if (robotName == "march3")
     {
-        this->value = march3;
+      this->value = march3;
     }
     else if (robotName == "testjoint_rotational")
     {
