+++ march_hardware_builder/test/TestHardwareBuilderCreatePowerDistributionBoard.cpp
@@ -7,8 +7,8 @@
 #include <march_hardware_builder/HardwareConfigExceptions.h>
 #include <march_hardware_builder/HardwareBuilder.h>
 
-using ::testing::Return;
 using ::testing::AtLeast;
+using ::testing::Return;
 
 class PowerDistributionBoardTest : public ::testing::Test
 {
