+++ march_hardware_interface/include/march_hardware_interface/PowerNetType.h
@@ -5,37 +5,69 @@
 
 #include <ros/package.h>
 
-class PowerNetType {
+class PowerNetType
+{
 public:
-  enum Value : int {  undefined = 0, high_voltage = 1, low_voltage = 2 };
+  enum Value : int
+  {
+    undefined = 0,
+    high_voltage = 1,
+    low_voltage = 2
+  };
 
-  PowerNetType() { value = undefined; }
+  PowerNetType()
+  {
+    value = undefined;
+  }
 
-  explicit PowerNetType(const std::string &name) {
-    if (name == "high_voltage") {
+  explicit PowerNetType(const std::string& name)
+  {
+    if (name == "high_voltage")
+    {
       this->value = high_voltage;
-    } else if (name == "low_voltage") {
+    }
+    else if (name == "low_voltage")
+    {
       this->value = low_voltage;
-    } else {
+    }
+    else
+    {
       ROS_ASSERT_MSG(false, "Unknown power net type %s", name.c_str());
       this->value = undefined;
     }
   }
 
-  bool operator==(PowerNetType a) const { return value == a.value; }
-  bool operator!=(PowerNetType a) const { return value != a.value; }
+  bool operator==(PowerNetType a) const
+  {
+    return value == a.value;
+  }
+  bool operator!=(PowerNetType a) const
+  {
+    return value != a.value;
+  }
 
-  bool operator==(int a) const { return value == a; }
-  bool operator!=(int a) const { return value != a; }
+  bool operator==(int a) const
+  {
+    return value == a;
+  }
+  bool operator!=(int a) const
+  {
+    return value != a;
+  }
 
   /** @brief Override stream operator for clean printing */
-  friend ::std::ostream &operator<<(std::ostream &os,
-                                    const PowerNetType &powerNetType) {
-    if (powerNetType.value == high_voltage) {
+  friend ::std::ostream& operator<<(std::ostream& os, const PowerNetType& powerNetType)
+  {
+    if (powerNetType.value == high_voltage)
+    {
       return os << "PowerNetType(type:HighVoltage)";
-    } else if (powerNetType.value == low_voltage) {
+    }
+    else if (powerNetType.value == low_voltage)
+    {
       return os << "PowerNetType(type:LowVoltage)";
-    } else {
+    }
+    else
+    {
       return os << "PowerNetType(type:Undefined)";
     }
   }
@@ -44,4 +76,4 @@ private:
   Value value;
 };
 
-#endif // MARCH_IV_POWERNETTYPE_H
+#endif  // MARCH_IV_POWERNETTYPE_H
