+++ march_hardware_interface/include/march_hardware_interface/march_hardware.h
@@ -15,9 +15,11 @@
 #include <march_hardware_interface/PowerNetType.h>
 #include <ros/ros.h>
 
-namespace march_hardware_interface {
+namespace march_hardware_interface
+{
 /// \brief Hardware interface for a robot
-class MarchHardware : public hardware_interface::RobotHW {
+class MarchHardware : public hardware_interface::RobotHW
+{
 protected:
   // Interfaces
   hardware_interface::JointStateInterface joint_state_interface_;
@@ -25,22 +27,15 @@ protected:
   hardware_interface::VelocityJointInterface velocity_joint_interface_;
   hardware_interface::EffortJointInterface effort_joint_interface_;
 
-  march_hardware_interface::MarchTemperatureSensorInterface
-      march_temperature_interface;
+  march_hardware_interface::MarchTemperatureSensorInterface march_temperature_interface;
   march_hardware_interface::MarchPdbStateInterface march_pdb_interface;
 
-  joint_limits_interface::EffortJointSaturationInterface
-      effort_joint_saturation_interface_;
-  joint_limits_interface::EffortJointSoftLimitsInterface
-      effort_joint_limits_interface_;
-  joint_limits_interface::PositionJointSaturationInterface
-      position_joint_saturation_interface_;
-  joint_limits_interface::PositionJointSoftLimitsInterface
-      position_joint_limits_interface_;
-  joint_limits_interface::VelocityJointSaturationInterface
-      velocity_joint_saturation_interface_;
-  joint_limits_interface::VelocityJointSoftLimitsInterface
-      velocity_joint_limits_interface_;
+  joint_limits_interface::EffortJointSaturationInterface effort_joint_saturation_interface_;
+  joint_limits_interface::EffortJointSoftLimitsInterface effort_joint_limits_interface_;
+  joint_limits_interface::PositionJointSaturationInterface position_joint_saturation_interface_;
+  joint_limits_interface::PositionJointSoftLimitsInterface position_joint_limits_interface_;
+  joint_limits_interface::VelocityJointSaturationInterface velocity_joint_saturation_interface_;
+  joint_limits_interface::VelocityJointSoftLimitsInterface velocity_joint_limits_interface_;
 
   // Custom or available transmissions
   // transmission_interface::RRBOTTransmission rrbot_trans_;
@@ -48,7 +43,7 @@ protected:
 
   // Shared memory
   int num_joints_;
-  int joint_mode_; // position, velocity, or effort
+  int joint_mode_;  // position, velocity, or effort
   std::vector<std::string> joint_names_;
   std::vector<int> joint_types_;
   std::vector<double> joint_position_;
@@ -61,7 +56,7 @@ protected:
 
   march4cpp::PowerDistributionBoard power_distribution_board_read_;
   bool master_shutdown_allowed_command = false;
-  //TODO(TIM) check if this is working::
+  // TODO(TIM) check if this is working::
   bool enable_high_voltage_command = true;
   PowerNetOnOffCommand power_net_on_off_command_;
 
@@ -72,8 +67,8 @@ protected:
   std::vector<double> joint_upper_limits_;
   std::vector<double> joint_effort_limits_;
 
-}; // class
+};  // class
 
-} // namespace
+}  // namespace march_hardware_interface
 
 #endif
