+++ march_hardware_interface/include/march_hardware_interface/march_temperature_sensor_interface.h
@@ -11,14 +11,11 @@ namespace march_hardware_interface
 class MarchTemperatureSensorHandle
 {
 public:
-  MarchTemperatureSensorHandle(
-      const std::string& name,      ///< The name of joint
-      const double* temperature,    ///< A pointer to the storage of the temperature value in degrees celsius.
-      const double* variance  ///< A pointer to the storage of the temperature covariance.
-      )
-    : name_(name)
-    , temperature_(temperature)
-    , variance(variance)
+  MarchTemperatureSensorHandle(const std::string& name,    ///< The name of joint
+                               const double* temperature,  ///< A pointer of the temperature in degrees celsius.
+                               const double* variance      ///< A pointer to the storage of the temperature covariance.
+                               )
+    : name_(name), temperature_(temperature), variance(variance)
   {
   }
 
@@ -47,6 +44,6 @@ private:
 class MarchTemperatureSensorInterface : public hardware_interface::HardwareResourceManager<MarchTemperatureSensorHandle>
 {
 };
-}
+}  // namespace march_hardware_interface
 
 #endif  // HARDWARE_INTERFACE_MARCH_TEMPERATURE_SENSOR_INTERFACE_H
